config/requirements.txt
# requirements.txt (runtime)
Django==4.2.6
graphene-django==3.2.0
django-graphql-jwt==0.3.4
psycopg-binary==3.1.10    # consider psycopg3 for Django 4.2+
channels==4.3.0
channels-redis==4.1.0
django-filter==23.5
python-decouple==3.8
Pillow==10.0